# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Main","description":"Function to load homepage.","operationId":"main__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/health":{"get":{"tags":["healthcheck"],"summary":"Perform a Health Check","description":"Function returning liveness info.","operationId":"healthcheck_health_get","responses":{"200":{"description":"Return HTTP Status Code 200 (OK)","content":{"application/json":{"schema":{}}}}}}},"/collect/":{"get":{"summary":"Collect Data","description":"Function to collect data.","operationId":"collect_data_collect__get","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analyze/{sentence}":{"get":{"summary":"Analyze Data","description":"Function to analyze sentence.","operationId":"analyze_data_analyze__sentence__get","parameters":[{"name":"sentence","in":"path","required":true,"schema":{"type":"string","title":"Sentence"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/db/{dbname}":{"get":{"summary":"Read Db","description":"Function to create db.","operationId":"read_db_db__dbname__get","parameters":[{"name":"dbname","in":"path","required":true,"schema":{"type":"string","title":"Dbname"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}